Refactor: patches as standalone files
